First Project 10-15-22
Webpage code created via https://www.freecodecamp.org. This lesson re-introduced me to a new way of critical thinking as the lesson gave you the instructions but sometimes the task was not as easy as written. The ability to critically think and break down tasks in order to solve a large problem is key in this profession and I had to use different resources on some occasions to answer questions. This is a good starting point for me and I am excited to see where I go.